This is the desktop client for oppster, which talks to the oserver. It was originally developed against AIR SDK 1.5.3 as a very QaD (Quick and Dirty) prototype.
I've very quickly ported it to the latest AIR SDK (3.3.0) with just enough changes to get it working.
The original checkin history was lost due to a need to sanitise the code and comments.
oclient really needs to be completely rewritten properly with modern idioms, so please do not think of this code as the right way of doing anything in AIR!